% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{decode_osrm_geom}
\alias{decode_osrm_geom}
\title{decode_osrm_geom}
\usage{
decode_osrm_geom(encoded)
}
\arguments{
\item{encoded}{}
}
\value{
geometry
}
\description{
decode_osrm_geom
}

