% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encoded_string_api_4}
\alias{encoded_string_api_4}
\title{encoded_string_api_4:
An encoded route to illustrate the 'osrmr::decode_geom()' function. After
decoding all points on the route are available as wgs84 coordinates. Decoding
varies on the API version of OSRMR. This version is decoded using API v4.}
\format{
A string containing an encoded polyline
}
\usage{
encoded_string_api_4
}
\description{
encoded_string_api_4:
An encoded route to illustrate the 'osrmr::decode_geom()' function. After
decoding all points on the route are available as wgs84 coordinates. Decoding
varies on the API version of OSRMR. This version is decoded using API v4.
}
\keyword{datasets}
