% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{make_request}
\alias{make_request}
\title{Run one server request for OSRM (online- or localhost)}
\usage{
make_request(request, t_max)
}
\arguments{
\item{request}{A character}

\item{t_max}{An integer, t_max after which an warning/error is thrown.}
}
\value{
A list. The dimension of the list depends on the request and wether the server reacted
properly or not.
}
\description{
In order to fail gracefully, this function handles errors and warnings if the asked
server (online- or localhost) doesn't work properly. In this case the error message is
returned and connections are closed using base::closeAllConnections().
}
\details{
If the asked server doesn't react within "t_max" seconds, a warning is thrown using
R.utils::withTimeout(..., t_max = t_max). t_max is initially set to 1sec.
}
