% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{quit_server}
\alias{quit_server}
\title{Quit local OSRM server}
\usage{
quit_server()
}
\description{
quit_server() quits your local OSRM server by using a taskkill shell command (depending on
your OS).
}
\examples{
\dontrun{
osrmr::quit_server()
# NULL}
}
