% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_address}
\alias{server_address}
\title{server_address}
\usage{
server_address(use_localhost)
}
\arguments{
\item{use_localhost}{A binary}
}
\value{
address The address of a OSRM server
}
\description{
server_address
}
\examples{
osrmr:::server_address(T)
# "http://localhost:5000"
osrmr:::server_address(F)
# "http://router.project-osrm.org"
}

