% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Calculate a distance matrix for given coordinates}
\usage{
table(coordinates, localhost, t_max = 1)
}
\arguments{
\item{coordinates}{A data.frame with 3 columns: id, lat, lng.
id can be numeric or a character
Lat and lng are coordinates in WGS84-format - numeric.
(-90 < lat < 90)
(-180 < lng <180)}

\item{localhost}{A logical(TRUE-localhost is used, FALSE=onlinehost is used)}

\item{t_max}{An integer, maximal time in seconds to wait for answer from server -
see function osrmr:::make_request()}
}
\value{
A matrix with the duration of all pairs of supplied coordinates
(Entry (i,j) gives the duration from point i to point j)
}
\description{
Given a set of coordinates, this function uses every coordinate once as start
and destination. For further details see:
http://project-osrm.org/docs/v5.5.1/api/#table-service
}
\examples{

coordinates <- data.frame(id=c("z", "b", "c"),
lat=c(47.4,47.5,47.6),
lng=c(9.04,9.04,9.04))
\dontrun{
# example with onlinehost
osrmr::table(coordinates,T)

# example with localhost
Sys.setenv("OSRM_PATH"="C:/OSRM_API5")
osrmr::run_server("switzerland-latest.osrm")
osrmr::table(coordinates,T)
osrmr::quit_server()
Sys.unsetenv("OSRM_PATH")}
}
