% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{trip}
\alias{trip}
\title{solves the Traveling Salesman Problem}
\usage{
trip(coordinates, localhost)
}
\arguments{
\item{coordinates}{A date.frame which contains 3 column: id, lat, lng
lat and lng are coordinates with the following notation:
(-90 < lat < 90)
(-180 < lng <180)}

\item{localhost}{A logical(TRUE-localhost is used, FALSE=onlinehost is used)}
}
\value{
A numeric giving the trip-index for every input-coordinates
}
\description{
The trip function solves the Traveling Salesman Problem. The returned path does not have to be the fastest path.
As TSP is NP-hard it only returns an approximation.
}
\examples{
coordinates <- data.frame(id=1:3,
lat=c(47.4623349064579,47.46229863897051,47.462226103920706),
lng=c(9.042273759841919,9.042563438415527,9.042906761169434))

trip(coordinates,F)
trip(coordinates,T)
}
