% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viaroute.R
\name{viaroute}
\alias{viaroute}
\title{Calculate route from source to destination (both given in WGS84). osrm chooses
the nearest point which can be accessed with a car. Be careful: Not all
combinations of api_version and localhost are possible}
\usage{
viaroute(lat1, lng1, lat2, lng2, instructions, api_version, localhost)
}
\arguments{
\item{lat1, }{A number}

\item{lng1, }{A number}

\item{lat2, }{A number}

\item{lng2, }{A number}

\item{instructions, }{A boolean for API V5: T = details are returned}

\item{api_version, }{A character}

\item{localhost, }{A boolean. T = local server used}
}
\value{
route
}
\description{
Calculate route from source to destination (both given in WGS84). osrm chooses
the nearest point which can be accessed with a car. Be careful: Not all
combinations of api_version and localhost are possible
}
\examples{
osrmr:::viaroute_api_v4(47,9,48,10,F, osrmr:::server_address(T))
# [1] 8250

osrmr:::viaroute_api_v5(47,9,48,10,F, osrmr:::server_address(F))
# [1] 8828.1
}

