% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viaroute.R
\name{viaroute_api_v5}
\alias{viaroute_api_v5}
\title{viaroute_api_v5}
\usage{
viaroute_api_v5(lat1, lng1, lat2, lng2, instructions, address)
}
\arguments{
\item{address}{}
}
\value{

}
\description{
viaroute_api_v5
}
\examples{
osrmr:::viaroute_api_v5(47,9,48,10,F, osrmr:::server_address(F))
# [1] 8828.1
}
