% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{decode_geom}
\alias{decode_geom}
\title{Transform encoded polylines to lat-lng list. Code from
http://stackoverflow.com/questions/32476218/how-to-decode-encoded-polylines-
from-osrm-and-plotting-route-geometry}
\usage{
decode_geom(encoded, api_version)
}
\arguments{
\item{encoded}{}
}
\value{
data.frame with lat and lng
}
\description{
Transform encoded polylines to lat-lng list. Code from
http://stackoverflow.com/questions/32476218/how-to-decode-encoded-polylines-
from-osrm-and-plotting-route-geometry
}
