\name{encoded_string_api_5}
\alias{encoded_string_api_5}
\docType{data}
\title{encoded test-dataset API 5}
\description{
  This is a test-dataset generated with the OSRM API 5 in order to test decode_geom(). The
  data contains the encoded polyline6 from (lat1 = 47.1 / lng1 = 8.1) to (lat2 = 46.9 /
  lng2 = 8.3) as character. The accuracy of this data is 1e-6
  }
\usage{data("encoded_string_api_5")}
\format{A character containing the encoded polyline from 47.1/8.1 to 46.9/8.3}
\source{http://router.project-osrm.org/route/v1/driving/8.1,47.1;8.3,46.9?steps=false&geometries=polyline6}
\keyword{datasets}
