% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{make_request}
\alias{make_request}
\title{Run one server request for OSRM (online- or localhost)}
\usage{
make_request(request)
}
\arguments{
\item{request}{A character}
}
\value{
A list. The dimension of the list depends on the request and wether the server reacted
properly or not.
}
\description{
In order to fail gracefully, this function handles errors and warnings if the asked
server (online- or localhost) doesn't work properly. In this case the error message is
returned and connections are closed using base::closeAllConnections().
}
\details{
If the asked server doesn't react within 1 second, a warning is thrown using
R.utils::withTimeout(..., timeout = 1)
}
