% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{Match GPS points to the road network in the most plaubislbe way}
\usage{
match(coordinates, localhost, t_max = 1)
}
\arguments{
\item{coordinates}{A data.frame with 2 columns "lat", "lng", giving a set of WGS84 coordinates.
(-90 < lat < 90)
(-180 < lng <180)}

\item{localhost}{A logical(TRUE-localhost is used, FALSE=onlinehost is used)}

\item{t_max}{An integer, maximal time in seconds to wait for answer from server -
see function osrmr:::make_request()}
}
\value{
A data.frame with the new coordinates
}
\description{
Details see: http://project-osrm.org/docs/v5.5.1/api/#match-service
Large jumps in the timestamps (> 60s) or improbable transitions lead to trace
splits if a complete matching could not be found. The algorithm might not be
able to match all points. Outliers are removed if they can not be matched successfully.
}
\examples{
coordinates <- data.frame(lat=c(47.4623,47.4622,47.4622),
lng=c(9.0422,9.0425,9.0429))
\dontrun{
# example with onlinehost
osrmr:::match(coordinates,F)

# example with localhost
Sys.setenv("OSRM_PATH"="C:/OSRM_API5")
osrmr::run_server("switzerland-latest.osrm")
osrmr::match(coordinates,T)
osrmr::quit_server()
Sys.unsetenv("OSRM_PATH")}
}
