% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R, R/routing.R
\name{nearest}
\alias{nearest}
\title{cacluate the nearest position which can be accessed via car. Be careful: Not
all combinations of api_version and localhost are possible}
\usage{
nearest(lat, lng, api_version, localhost)

nearest(lat, lng, api_version, localhost)
}
\arguments{
\item{lat, }{A number}

\item{lng, }{A number}

\item{api_version, }{A character}

\item{localhost, }{A boolean. T = local server used}

\item{lat, }{A number}

\item{lng, }{A number}

\item{api_version, }{A character}

\item{localhost, }{A boolean. T = local server used}
}
\value{
nearest

nearest
}
\description{
cacluate the nearest position which can be accessed via car. Be careful: Not
all combinations of api_version and localhost are possible

cacluate the nearest position which can be accessed via car. Be careful: Not
all combinations of api_version and localhost are possible
}

