% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{nearest accessible position}
\usage{
nearest(lat, lng, api_version = 5, localhost = F)
}
\arguments{
\item{lat, }{A numeric (-90 < lat < 90)}

\item{lng, }{A numeric (-180 < lng < 180)}

\item{api_version, }{A numeric (either 4 or 5)}

\item{localhost, }{A logical (TRUE = localhost is used, FALSE = onlinehost is used)}
}
\value{
A data.frame with lat and lng
}
\description{
nearest() calculates the nearest position to the given coordinates which can be accessed by car.
The coordinate-standard is WGS84. Attention: The OSRM API v4 is only working locally, but
not with the 'OSRM' webserver.
}
\examples{
# osrmr::nearest(47,9, 5, F)
#        lat      lng
# 1 47.02872 9.006633

# osrmr::run_server("C:/OSRM_API5/", "switzerland-latest.osrm")
# osrmr:::nearest(47,9, 5, T)
#        lat      lng
# 1 47.02872 9.006633
# osrmr::quit_server()

# osrmr::run_server("C:/OSRM_API4/", "switzerland-latest.osrm")
# osrmr:::nearest(47,9, 4, T)
#        lat      lng
# 1 47.02891 9.006577
# osrmr::quit_server()
}
