% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{cacluate the nearest position which can be accessed via car. Be careful: Not
all combinations of api_version and localhost are possible}
\usage{
nearest(lat, lng, api_version = 5, localhost = F)
}
\arguments{
\item{lat, }{A number}

\item{lng, }{A number}

\item{api_version, }{A character}

\item{localhost, }{A boolean. T = local server used}
}
\value{
nearest
}
\description{
cacluate the nearest position which can be accessed via car. Be careful: Not
all combinations of api_version and localhost are possible
}
\examples{
osrmr:::nearest_api_v4(47,9, osrmr:::server_address(T))
#        lat      lng
# 1 47.02872 9.006633

osrmr:::nearest_api_v5(47,9, osrmr:::server_address(F))
#        lat      lng
# 1 47.02871 9.006633
}
