% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest_api_v4}
\alias{nearest_api_v4}
\title{nearest_api_v4}
\usage{
nearest_api_v4(lat, lng, address)
}
\arguments{
\item{address}{}
}
\value{
nearest A data.frame with lat and lng
}
\description{
nearest_api_v4
}
\examples{
osrmr:::nearest_api_v4(47,9, osrmr:::server_address(T))
#        lat      lng
# 1 47.02872 9.006633
}
