% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest_api_v5}
\alias{nearest_api_v5}
\title{nearest_api_v5}
\usage{
nearest_api_v5(lat, lng, address)
}
\arguments{
\item{address}{}
}
\value{
nearest A data.frame with lat and lng
}
\description{
nearest_api_v5
}
\examples{
osrmr:::nearest_api_v5(47,9, osrmr:::server_address(F))
#        lat      lng
# 1 47.02871 9.006633
}

