% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest_api_v5}
\alias{nearest_api_v5}
\title{nearest accessible position for OSRM API v5}
\usage{
nearest_api_v5(lat, lng, address)
}
\arguments{
\item{lat, }{A numeric (-90 < lat < 90)}

\item{lng, }{A numeric (-180 < lng < 180)}

\item{address, }{A character specifying the serveraddress (local or online)}
}
\value{
A data.frame with lat and lng
}
\description{
nearest_api_v5() calculates the nearest position to the given coordinates which can be accessed
by car with the OSRM API v5. The coordinate-standard is WGS84.
}
\examples{
# osrmr::run_server("C:/OSRM_API5/", "switzerland-latest.osrm")
# osrmr:::nearest_api_v5(47,9, osrmr:::server_address(T))
#        lat      lng
# 1 47.02872 9.006633
# osrmr:::nearest_api_v5(47,9, osrmr:::server_address(F))
# lat      lng
# 1 47.02872 9.006633
# osrmr::quit_server()
}
