% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmr.R
\name{prepare_map}
\alias{prepare_map}
\title{In case of a new map: OSRM prepare (Creating the Hierarchy, creating the graph, nodeset etc..)}
\usage{
prepare_map(osrm_path, plattform = "windows")
}
\arguments{
\item{osrm_path, }{A character}
}
\description{
In case of a new map: OSRM prepare (Creating the Hierarchy, creating the graph, nodeset etc..)
}

