% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{run_server}
\alias{run_server}
\title{Start OSRM with map "switzerland-latest.osrm"}
\usage{
run_server(osrm_path)
}
\arguments{
\item{osrm_path}{A string containing the path where OSRM is installed}
}
\value{
error_code A character
}
\description{
Start OSRM with map "switzerland-latest.osrm"
}
\examples{
# Start Server on SCNs Mac
osrmr::run_server("/Users/adrianschmid/Documents/OSRM/osrm-backend/build/")
}
