% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{run_server}
\alias{run_server}
\title{Start local OSRM server}
\usage{
run_server(osrm_path, map_name)
}
\arguments{
\item{osrm_path}{A character containing the directory of the local OSRM-engine}

\item{map_name}{A character (name of your pre-built map - e.g. "switzerland-latest.osrm")}
}
\value{
error_code A character
}
\description{
run_server() starts your local OSRM server by using a shell command (depending on
your OS). A local (pre-built) version of the OSRM-engine must be on your device.
(https://github.com/Project-OSRM/osrm-backend/wiki/Building-OSRM).
}
\examples{
\dontrun{
base::Sys.setenv("OSRM_PATH_API_5"="C:/OSRM_API5")
osrmr::run_server(base::Sys.getenv("OSRM_PATH_API_5"), "switzerland-latest.osrm")
base::Sys.unsetenv("OSRM_PATH_API_5")
base::Sys.setenv("OSRM_PATH_API_4"="C:/OSRM_API4")
osrmr::run_server(base::Sys.getenv("OSRM_PATH_API_4"), "switzerland-latest.osrm")
base::Sys.unsetenv("OSRM_PATH_API_4")}
}
