% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_address}
\alias{server_address}
\title{Set server address}
\usage{
server_address(use_localhost)
}
\arguments{
\item{use_localhost}{A logical, indicating whether to use the localhost or not.}
}
\value{
character, the address of an OSRM server
}
\description{
server_address() returns an address depending on the choice of a localhost or webserver.
This address is used as a part of a OSRM server-request.
}
\examples{
osrmr:::server_address(TRUE)
# "http://localhost:5000"
osrmr:::server_address(FALSE)
# "http://router.project-osrm.org"
}
