% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_address}
\alias{server_address}
\title{server_address() returns the URL address of the OSRM localhost or OSRM webserver,
depending on the value of the variable 'use_localhost'.
This is an internal function. The address is used as a part of a OSRM server-request.}
\usage{
server_address(use_localhost)
}
\arguments{
\item{use_localhost}{A logical, indicating whether to use the localhost or not.}
}
\value{
character, the address of an OSRM server
}
\description{
server_address() returns the URL address of the OSRM localhost or OSRM webserver,
depending on the value of the variable 'use_localhost'.
This is an internal function. The address is used as a part of a OSRM server-request.
}
\examples{
osrmr:::server_address(TRUE)
# [1] "http://localhost:5000"
osrmr:::server_address(FALSE)
# [1] "http://router.project-osrm.org"
}
