% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{trip}
\alias{trip}
\title{Find a solution for the TSP (Traveling Salesman Problem)}
\usage{
trip(coordinates, localhost, t_max = 1)
}
\arguments{
\item{coordinates}{A data.frame with 2 columns: lat, lng.
lat and lng are coordinates in WGS84-format - numeric.
(-90 < lat < 90)
(-180 < lng <180)}

\item{localhost}{A logical(TRUE-localhost is used, FALSE=onlinehost is used)}

\item{t_max}{An integer, maximal time in seconds to wait for answer from server -
see function osrmr:::make_request()}
}
\value{
A numeric vector giving the trip-index for every input-coordinate
}
\description{
The trip function solves the Traveling Salesman Problem using a greedy heuristic
(farthest-insertion algorithm). The returned path does not have to be the fastest
path as TSP is NP-hard it is only an approximation. Note that if the input coordinates
can not be joined by a single trip (e.g. the coordinates are on several disconnected
islands) multiple trips for each connected component are returned.
For further details see:
http://project-osrm.org/docs/v5.5.1/api/#trip-service
}
\examples{
coordinates <- data.frame(lat=c(47.4623,47.4622,47.4622),lng=c(9.0422,9.0425,9.0429))

\dontrun{
# example with onlinehost
osrmr:::trip(coordinates,F)

# example with localhost
Sys.setenv("OSRM_PATH"="C:/OSRM_API5")
osrmr::run_server("switzerland-latest.osrm")
osrmr::trip(coordinates,T)
osrmr::quit_server()
Sys.unsetenv("OSRM_PATH")}
}
