% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viaroute.R
\name{viaroute}
\alias{viaroute}
\title{Calculate travel time or full information of a route}
\usage{
viaroute(
  lat1,
  lng1,
  lat2,
  lng2,
  localhost,
  instructions = FALSE,
  timeout = 0.001
)
}
\arguments{
\item{lat1}{A numeric (-90 < lat1 < 90) -> start-destination}

\item{lng1}{A numeric (-180 < lng1 < 180) -> start-destination}

\item{lat2}{A numeric (-90 < lat2 < 90) -> end-destination}

\item{lng2}{A numeric (-180 < lng2 < 180) -> end-destination}

\item{localhost}{A logical (TRUE = localhost is used, FALSE = onlinehost is used)}

\item{instructions}{A logical. If FALSE, only the traveltime (in seconds,
as numeric) will be returned.
 If TRUE, the whole route is returned as list}

\item{timeout}{}
}
\value{
a numeric or a list (depending on instructions)
}
\description{
For a given start- and end-destination, route() calculates route informations
using OSRM. OSRM chooses the nearest point which can be accessed by car for
the start- and end-destination. The coordinate-standard is WGS84.
}
\examples{
# direct examples of the online API
\dontrun{
link <- "http://router.project-osrm.org/route/v1/driving/8.1,47.1;8.3,46.9?steps=false"
a <- rjson::fromJSON(file = link)

# example with onlinehost
osrmr::viaroute(47.1, 8.1, 46.9, 8.3, FALSE, FALSE)
osrmr::viaroute(47.1, 8.1, 46.9, 8.3, FALSE, TRUE)

# example with localhost
Sys.setenv("OSRM_PATH"="C:/OSRM_API5")
osrmr::run_server("switzerland-latest.osrm")
osrmr::viaroute(47.1, 8.1, 46.9, 8.3, TRUE, FALSE)
osrmr::viaroute(47.1, 8.1, 46.9, 8.3, TRUE, TRUE)
osrmr::quit_server()
Sys.unsetenv("OSRM_PATH")}
}
